+++ main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"net/http"
+	"strconv"
 
 	"github.com/gin-gonic/gin"
 	"github.com/rabdill/monkeysim/monkey"
@@ -15,6 +16,8 @@ func main() {
 
 	router.GET("/monkeys", info)
 	router.POST("/monkeys", addMonkey)
+
+	router.PATCH("/monkeys/:id/stand", stand)
 
 	go monkey.KickOffSim()
 
@@ -32,4 +35,19 @@ func addMonkey(c *gin.Context) {
 		return
 	}
 	c.JSON(201, monkey)
+}
+
+func stand(c *gin.Context) {
+	id, err := strconv.Atoi(c.Param("id"))
+	if err != nil {
+		c.JSON(500, err)
+		return
+	}
+
+	err = monkey.StandUp(id)
+	if err != nil {
+		c.JSON(500, err)
+		return
+	}
+	c.Status(202)
 }