+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/rabdill/monkeysim/compare/v0.2.0...HEAD)
-- none
+- Monkeys can be removed from their seats
 
 ## 0.2.0 - 2017-12-12
 - **Converted command-line interface into web UI**